+++ openai-client/src/commonTest/kotlin/com/aallam/openai/client/misc/TestException.kt
@@ -4,16 +4,16 @@ import com.aallam.openai.api.exception.AuthenticationException
 import com.aallam.openai.api.exception.OpenAIAPIException
 import com.aallam.openai.api.model.ModelId
 import com.aallam.openai.client.OpenAI
-import kotlinx.coroutines.test.runTest
+import com.aallam.openai.client.TestOpenAI
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertIs
 import kotlin.test.assertTrue
 
-class TestException {
+class TestException : TestOpenAI() {
 
     @Test
-    fun apiError() = runTest {
+    fun apiError() = test {
         val openAI = OpenAI("sk-123")
         val model = runCatching {
             openAI.model(ModelId("davinci"))